Links deployment confidence dashboard on deploy creation msg
While deploying it is important that we check the vitals on our services to ensure
nothing we've deployed is breaking.

This commit links to our datadog deployment confidence dashboard to allow us to do just that.
There is still a small caveat that the dashboard linked isn't yet flexible enough to allow switching
between environments (production|staging-tfe), which is something we are looking to tackle in a
follow-up commit.